Back/requirements.txt
# =============================================
# Instagram Scraper — Dependencias del Backend
# =============================================

# --- Core de scraping (API privada de Instagram) ---
instagrapi>=2.1.0          # Cliente API privada IG (sin Selenium/BS4)

# --- Variables de entorno ---
python-dotenv>=1.0.0       # Carga .env automáticamente

# --- HTTP y utilerías base ---
requests>=2.31.0           # HTTP client (dependencia de instagrapi)
Pillow>=10.0.0             # Procesamiento de imágenes (requerido por instagrapi)

# --- Validación de datos ---
pydantic>=2.0.0            # Validación de modelos (soporte para DDD)
